refactor (LottoGenerator) : modify generator => 인자에 로또 넘버를 넘겨받도록 변경 및 LottoTicketManualGeneratorTest LottoTicketGeneratorTest로 이동
