+++ src/main/java/lotto/domain/ticket/LottoTicketGenerator.java
@@ -1,7 +1,13 @@
 package lotto.domain.ticket;
 
+import lotto.domain.LottoNumber;
+
+import java.util.List;
+
 @FunctionalInterface
 public interface LottoTicketGenerator {
 
-    LottoTicket generate();
+    int LOTTO_GENERATE_MAX_COUNT = 6;
+
+    LottoTicket generate(List<LottoNumber> lottoNumbers);
 }
